+++ src/traits.rs
@@ -131,7 +131,7 @@ pub trait Step {
 ///
 /// The `Resources` type parameter defines the hardware resources required for
 /// motion control.
-pub trait EnableMotionControl<Resources, const TIMER_HZ: u32> {
+pub trait EnableMotionControl<Resources, const TIMER_HZ: u64> {
     /// The type of the driver after motion control has been enabled
     type WithMotionControl: MotionControl;
